Fixed Board, Preliminary support for Dashboards, etc
Lots of little fixes in this one, mostly in the Board view. I also added preliminary support for most of the new Dashboards feature, though there is still a little bit that I can't quite get to yet.
I fixed a white placeholder that you see during loading, which is cool! There are still some more of those lurking, but they're hard for me to nail down because it's usually loaded by the time I get in there to inspect the code. Oh well! I have a note to just look in the source CSS for "loading" or "placeholder" to see what I can dig up.